Fixes for llvm-pipe

+++ src/editor/res/shaders/circular_brushed_metal.frag
@@ -45,8 +45,8 @@ void main() {
         1u, 1u, 2u, 2u, 1u, 1u, 2u, 2u
     );
     const uvec2 dither_pos   = uvec2(
-        uint(gl_FragCoord.x) + odd_bits[gl_SampleID & 31u],
-        uint(gl_FragCoord.y) + even_bits[gl_SampleID & 31u]
+        uint(gl_FragCoord.x) + odd_bits [uint(gl_SampleID) & 31u],
+        uint(gl_FragCoord.y) + even_bits[uint(gl_SampleID) & 31u]
     ) % 16u;
     const uint  dither_index = (dither_pos.y * 16u + dither_pos.x + gl_SampleID * 211u + v_draw_id * 241u) & 255u;
     const float dither_value = bayer_matrix[dither_index];
